% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{cluster_by_amplitude}
\alias{cluster_by_amplitude}
\title{Cluster the time series according to the y-axis using k-means algorithm.}
\usage{
cluster_by_amplitude(ts, num.cl)
}
\arguments{
\item{ts}{Input vector or a time-series object to be clustered.}

\item{num.cl}{Number of clusters (e.g. estimated by \link{estimate_clusters} function)}
}
\value{
A structure \code{list(data, num.cl, cfit, thdata, abserror, MAE)} where:
\describe{
 \item{data}{Copied input parameter \code{ts}.}
 \item{num.cl}{Copied input parameter \code{num.cl}.}
 \item{cfit}{Result from the \link{kmeans} function.}
 \item{thdata}{The new time-series object with values replaced with the cluster centers.}
 \item{abserror}{A vector of absolute errors, i.e., \code{abs(data - thdata)}}
 \item{MAE}{Mean of all absolute errors}
}
}
\description{
Cluster the time series according to the y-axis using k-means algorithm.
}

