% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compute_emits}
\alias{compute_emits}
\title{Simulate generators' activity.}
\usage{
compute_emits(D)
}
\arguments{
\item{D}{Matrix of coeficients from the MSD, e.g. \code{dts$coef}}
}
\value{
Matrix of coeficients
}
\description{
Simulate generators' activity.
}

