% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{compute_stairs}
\alias{compute_stairs}
\title{Helper function - TODO comment}
\usage{
compute_stairs(ts)
}
\arguments{
\item{ts}{Input time series}
}
\value{
\code{list(x,y)}
}
\description{
Helper function - TODO comment
}

