% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decompose}
\alias{decompose}
\title{Multi-scale decomposition of time series.
(most important function in this package)}
\usage{
decompose(ts, cutoff.fraction = 0.004)
}
\arguments{
\item{ts}{Input vector which represents regularized time-series.}

\item{cutoff.fraction}{How strongly to optimize the coeficients.}
}
\value{
A structure \code{list(length, maxscale, coef)}
}
\description{
The result is a structure \code{list(length, maxscale, coef)} where:
\describe{
 \item{length}{Size of the input vector (number of samples)}
 \item{maxscale}{What is the maximum scale produced by the decomposition.}
 \item{coef}{
   The output matrix containing the decomposition coeficients.
   Rows represent scales (row 1 represents scale 0).
   Columns represent time (sample with index same as column number).
 }
}
}

