% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw_image_legend}
\alias{draw_image_legend}
\title{Helper function to plot a nice legend inside an image.}
\usage{
draw_image_legend(text, nrow, ncol)
}
\arguments{
\item{text}{Text to be plotted}

\item{nrow}{Y-position of the text}

\item{ncol}{X-position of the text}
}
\value{
see \link{legend}
}
\description{
Helper function to plot a nice legend inside an image.
}

