% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{estimate_clusters}
\alias{estimate_clusters}
\title{Estimate number of clusters.}
\usage{
estimate_clusters(ts, max.cl = 25, cutoff = 0.1)
}
\arguments{
\item{ts}{Input vector of numbers (preferably integers), can also be a 
Time-series object created using \code{ts()} function.}

\item{max.cl}{The maximum number of clusters that the algorithm evaluates.}

\item{cutoff}{Determines how the right number of clusters is chosen.}
}
\value{
The estimated number of clusters. (More clusters will have
  \code{sum(witness) < cutoff})
}
\description{
Estimate number of clusters.
}

