% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{get_interval_count}
\alias{get_interval_count}
\title{Count non-empty intervals in a vector.}
\usage{
get_interval_count(vec)
}
\arguments{
\item{vec}{Input vector of numbers.}
}
\value{
The number of intervals found inside the vector.
}
\description{
Count non-empty intervals in a vector.
}

