% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{get_total_intervals}
\alias{get_total_intervals}
\title{Count non-empty intervals in the matrix of coeficients.}
\usage{
get_total_intervals(coef)
}
\arguments{
\item{coef}{Input matrix, where each row represent a single vector of coeficients.}
}
\value{
The number of intervals found.
}
\description{
Count non-empty intervals in the matrix of coeficients.
}

