% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decomposition.R
\name{optimize_intervals}
\alias{optimize_intervals}
\title{Optimization used as a part of the MSD algorithm.}
\usage{
optimize_intervals(x, cutoff.fraction = 0.004)
}
\arguments{
\item{x}{Input Vector}

\item{cutoff.fraction}{How strongly to optimize the coeficients.}
}
\value{
Optimized version of the input vector.
}
\description{
Optimization used as a part of the MSD algorithm.
}

