% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decomposition.R
\name{pad_power2}
\alias{pad_power2}
\title{Pad the signal with zeroes to the nearest higher power of two.}
\usage{
pad_power2(ts)
}
\arguments{
\item{ts}{Input vector of numbers (preferably integers)}
}
\value{
New vector that is padded with zeroes
}
\description{
Pad the signal with zeroes to the nearest higher power of two.
}

