% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot spectrum of the decomposition.}
\usage{
plot_spectrum(D, show.emits = TRUE, show.spectrum = TRUE,
  show.contour = TRUE, show.istats = TRUE, show.hmlegend = TRUE)
}
\arguments{
\item{D}{The input matrix with decomposed coeficients.}

\item{show.emits}{If TRUE, renders a separate image depicting places where
packets are emitted.}

\item{show.spectrum}{If TRUE, renders a nice image showing the spectrum based.
The spectrum is constructed from the coeficients.}

\item{show.contour}{If TRUE, contours are rendered within the spectrum.}

\item{show.istats}{If TRUE, renders additional information about intervals.}

\item{show.hmlegend}{Renders the "heatmap legend" in the spectrogram.}

\item{show.cwt}{If TRUE, renders a spectrum using continuous wavelet transformation.}
}
\description{
Plot spectrum of the decomposition.
}
\examples{
traffic <- read_traffic(filename)
 D <- decompose(traffic)
 plot_spectrum(D)
}

