% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_traffic}
\alias{plot_traffic}
\title{Plotting network traffic as a single time series or as a comparison of two time series.}
\usage{
plot_traffic(ts, newts = c(0), area.color = "black")
}
\arguments{
\item{ts}{First time series}

\item{newts}{Optional time series for comparison}

\item{area.color}{Color for the area under the curve.}
}
\description{
Plotting network traffic as a single time series or as a comparison of two time series.
}

