% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_traffic_densities}
\alias{plot_traffic_densities}
\title{Plotting density of a traffic time series.}
\usage{
plot_traffic_densities(ts.list, area.colors = index(ts.list))
}
\arguments{
\item{ts.list}{List of time series}

\item{area.colors}{List of colors}
}
\description{
Plotting density of a traffic time series.
}

