% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pretty_axis}
\alias{pretty_axis}
\title{Create pretty axis using pretty labels}
\usage{
pretty_axis(side, v, howmany = length(v), project.on = c(0, 1))
}
\arguments{
\item{side}{See the "size" parameter in "axis()" function.}

\item{v}{See function \link{pretty_labels}}

\item{howmany}{See function \link{pretty_labels}}

\item{project.on}{See function \link{pretty_labels}}
}
\description{
Create pretty axis using pretty labels
}

