% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pretty_labels}
\alias{pretty_labels}
\title{Create pretty labels.}
\usage{
pretty_labels(v, howmany = length(v), project.on = c(0, 1))
}
\arguments{
\item{v}{Vector used as a range from which we draw our labels.}

\item{howmany}{Howmany labels are preferred to be drawn on the axis.}

\item{project.on}{The original interval of labels default [0,1] to which we
want to project our newly created lables.}
}
\value{
A list of two elements \code{list(at, labels)}.
}
\description{
Create pretty labels.
}
\seealso{
\link{axis} function for the sematics of the "at" and "labels" elements.
}

