% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{print_stats}
\alias{print_stats}
\title{Print statistics about the multiscale decomposition}
\usage{
print_stats(D)
}
\arguments{
\item{D}{Matrix containing coeficients.}
}
\description{
Print statistics about the multiscale decomposition
}

