% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decomposition.R
\name{read_traffic}
\alias{read_traffic}
\title{Reads traffic data from a CSV file.}
\usage{
read_traffic(filename, use.power2len = TRUE, use.expfactor = 1,
  aggregation_function = sum)
}
\arguments{
\item{filename}{The traffic data will be read from this file.}

\item{use.power2len}{Length padded with zeroes to power of 2.}

\item{use.expfactor}{How should the time be stretched (for debugging)}
}
\value{
Time-series object which contains the regularized time series.
}
\description{
The input file should be a CSV files with the first row containing column names.
The file should contain at least two columns: Time, Length
}

