% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{reconstruct_traffic}
\alias{reconstruct_traffic}
\title{Simulate the reconstruction of the original signal.}
\usage{
reconstruct_traffic(D)
}
\arguments{
\item{D}{Either the matrix of coeficients or a \code{list(coef)} that contains the
matrix as a \code{coef}.}
}
\value{
Vector representing the orignal signal in a form of regularized time
  series.
}
\description{
Simulate the reconstruction of the original signal by using the matrix of 
decomposed coeficients. The matrix needs to be filtered in order to simulate 
the "generators" emitting the packets. Let's assume a generator on
\code{scale=3}. It would emit a packet every 8 samples. However, the matrix
will contain 8 coeficients for every packet that should be emitted. Therefore
the input matrix needs to be filtered.
}

