% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficmsd-package.R
\docType{data}
\name{sample32}
\alias{sample32}
\title{Simple traffic example of 32 samples for debugging}
\format{An object of class \code{numeric} of length 32.}
\usage{
sample32
}
\description{
Simple traffic example of 32 samples for debugging
}
\keyword{datasets}

