% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{sample_linkcap}
\alias{sample_linkcap}
\title{Example of heat-up phase with large data transfer}
\format{\preformatted{ int [1:594] 3918 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
sample_linkcap
}
\description{
Irregular data transfer with a heat-up phase containing several small data
transfers, followed by a continuous transfer reaching the link capacity.
}
\keyword{datasets}

