% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficmsd-package.R
\docType{data}
\name{sample_regsignal}
\alias{sample_regsignal}
\title{Example of regular signal}
\format{An object of class \code{integer} of length 597.}
\usage{
sample_regsignal
}
\description{
Example of regular signal
}
\keyword{datasets}

