% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{sample_vidstream}
\alias{sample_vidstream}
\title{Video streaming traffic}
\format{\preformatted{ int [1:3584] 349 6036 1718 0 0 0 0 0 0 0 ...
}}
\usage{
sample_vidstream
}
\description{
Data transfer during 60 minutes of video streaming with occasional caching.
}
\keyword{datasets}

