% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficmsd-package.R
\docType{data}
\name{sample_withoutliers}
\alias{sample_withoutliers}
\title{Signal with several outliers.}
\format{An object of class \code{integer} of length 600.}
\usage{
sample_withoutliers
}
\description{
An example of multiple applications transferring data in parallel.
}
\keyword{datasets}

