% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{shade_color}
\alias{shade_color}
\title{Creates a shaded version of given colors.}
\usage{
shade_color(colors, percent)
}
\arguments{
\item{colors}{vector of any of the three kinds of R color specifications.}

\item{percent}{If \code{percent > 1} the colors will be lighter, if \code{percent < 1} then darker.}
}
\description{
Creates a shaded version of given colors.
}

