% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trafficmsd-package.R
\docType{data}
\name{sample32}
\alias{sample32}
\title{Simple traffic example of 32 samples for debugging}
\format{\preformatted{ num [1:32] 0 0 0 0 4 4 4 0 2 2 ...
}}
\usage{
sample32
}
\description{
Simple traffic example of 32 samples for debugging
}
\keyword{datasets}

