% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{sample_regsignal}
\alias{sample_regsignal}
\title{Example of regular signal}
\format{\preformatted{ int [1:597] 792 180 0 0 0 486 432 0 0 0 ...
}}
\usage{
sample_regsignal
}
\description{
Example of regular signal
}
\keyword{datasets}

