% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficmsd-package.R
\docType{data}
\name{sample_vidstream}
\alias{sample_vidstream}
\title{Video streaming traffic}
\format{An object of class \code{integer} of length 3584.}
\usage{
sample_vidstream
}
\description{
Data transfer during 60 minutes of video streaming with occasional caching.
}
\keyword{datasets}

