% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{sample_withoutliers}
\alias{sample_withoutliers}
\title{Signal with several outliers.}
\format{\preformatted{ int [1:600] 60 60 5472 60 0 0 0 0 60 0 ...
}}
\usage{
sample_withoutliers
}
\description{
An example of multiple applications transferring data in parallel.
}
\keyword{datasets}

